nsgfwin debugged against matlab
git-svn-id: https://svn.grrrr.org/projects/dev/nsgt@680 3d3ad2b0-3bb8-11dd-8da9-e1ab7d56d6ab

+++ src/util.py
@@ -8,7 +8,7 @@ import numpy as N
 from math import pi
 
 def hannwin(l):
-    r = N.arange(l)
+    r = N.arange(l,dtype=float)
     r *= pi*2./l
     r = N.cos(r)
     r += 1.
